Fixed warnings in semaphore code (following similar commit in readerwriterqueue)

+++ lightweightsemaphore.h
@@ -139,7 +139,7 @@ public:
 	{
 		mach_timespec_t ts;
 		ts.tv_sec = static_cast<unsigned int>(timeout_usecs / 1000000);
-		ts.tv_nsec = (timeout_usecs % 1000000) * 1000;
+		ts.tv_nsec = static_cast<int>((timeout_usecs % 1000000) * 1000);
 
 		// added in OSX 10.10: https://developer.apple.com/library/prerelease/mac/documentation/General/Reference/APIDiffsMacOSX10_10SeedDiff/modules/Darwin.html
 		kern_return_t rc = semaphore_timedwait(m_sema, ts);
@@ -175,7 +175,7 @@ public:
 	Semaphore(int initialCount = 0)
 	{
 		assert(initialCount >= 0);
-		int rc = sem_init(&m_sema, 0, initialCount);
+		int rc = sem_init(&m_sema, 0, static_cast<unsigned int>(initialCount));
 		assert(rc == 0);
 		(void)rc;
 	}
